+++ src/main/java/com/project/ets/controller/BatchController.java
@@ -1,7 +1,5 @@
 package com.project.ets.controller;
 
-import java.time.LocalDate;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -11,31 +9,19 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.project.ets.enums.UserRole;
 import com.project.ets.requstdto.BatchRequest;
 import com.project.ets.responsedto.BatchResponse;
-import com.project.ets.responsedto.UserResponse;
-import com.project.ets.security.RegistrationRequest;
 import com.project.ets.service.BatchService;
-import com.project.ets.service.UserService;
 import com.project.ets.util.AppResponseBuilder;
 import com.project.ets.util.ResponseStructure;
 
-import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 
 @RestController
 @AllArgsConstructor
-public class AdminController {
-	private UserService adminService;
-	private AppResponseBuilder responseBuilder;
+public class BatchController {
 	private BatchService batchService;
-	
-	@PostMapping("/users/admins/{role}")
-	public ResponseEntity<ResponseStructure<UserResponse>> saveAdmin(@RequestBody @Valid RegistrationRequest registrationRequest,@PathVariable UserRole role) {
-		UserResponse adminResponse=adminService.saveUser(registrationRequest,role);
-		return responseBuilder.success(HttpStatus.CREATED,"Admin created Successfully", adminResponse);
-	}
+	private AppResponseBuilder responseBuilder;
 	
 	@PostMapping("/admins/batches")
 	public ResponseEntity<ResponseStructure<BatchResponse>> saveBatch(@RequestBody BatchRequest request){
