+++ src/main/java/com/project/ets/entity/Batch.java
@@ -1,11 +1,9 @@
 package com.project.ets.entity;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
 
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import com.project.ets.config.GenerateSequenceId;
