+++ src/main/java/com/project/ets/requstdto/TrainerRequest.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import com.project.ets.enums.Subject;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,7 +12,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TrainerRequest extends UserRequest{
-	
+	@NotNull
 	private List<Subject> subjects;
 
 }
